Away from size_t. It's signed/unsigned on different platforms.
Use Vector types (WORD, DWORD, ...) were appropriate.

+++ src/Vector/BLF/AppText.h
@@ -60,7 +60,7 @@ public:
 
     virtual void read(AbstractFile & is);
     virtual void write(AbstractFile & os);
-    virtual size_t calculateObjectSize();
+    virtual DWORD calculateObjectSize();
 
     /** enumeration for source */
     enum class Source : DWORD {
