+++ src/Vector/BLF/FlexRayVFrStatus.h
@@ -50,7 +50,7 @@ public:
 
     virtual void read(AbstractFile & is);
     virtual void write(AbstractFile & os);
-    virtual size_t calculateObjectSize();
+    virtual DWORD calculateObjectSize();
 
     /**
      * @brief application channel
