+++ src/Vector/BLF/LinMessageDescriptor.h
@@ -43,7 +43,7 @@ public:
 
     virtual void read(AbstractFile & is);
     virtual void write(AbstractFile & os);
-    virtual size_t calculateObjectSize();
+    virtual DWORD calculateObjectSize();
 
     /**
      * @brief LIN Sub-Identifier - Supplier ID
