+++ src/Vector/BLF/WlanFrame.h
@@ -46,7 +46,7 @@ public:
 
     virtual void read(AbstractFile & is);
     virtual void write(AbstractFile & os);
-    virtual size_t calculateObjectSize();
+    virtual DWORD calculateObjectSize();
 
     /**
      * @brief application channel 1..n
